ändrade int till double för att lösa divisonsproblem

+++ ACMECalc/Form1.cs
@@ -19,36 +19,36 @@ namespace ACMECalc
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int i, j;
-            i = int.Parse(textBox1.Text);
-            j = int.Parse(textBox2.Text);
+            double i, j;
+            i = double.Parse(textBox1.Text);
+            j = double.Parse(textBox2.Text);
 
             listBox1.Items.Add($"{i} + {j} = {i + j}");
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int i, j;
-            i = int.Parse(textBox1.Text);
-            j = int.Parse(textBox2.Text);
+            double i, j;
+            i = double.Parse(textBox1.Text);
+            j = double.Parse(textBox2.Text);
 
             listBox1.Items.Add($"{i} - {j} = {i - j}");
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int i, j;
-            i = int.Parse(textBox1.Text);
-            j = int.Parse(textBox2.Text);
+            double i, j;
+            i = double.Parse(textBox1.Text);
+            j = double.Parse(textBox2.Text);
 
             listBox1.Items.Add($"{i} / {j} = {i / j}");
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int i, j;
-            i = int.Parse(textBox1.Text);
-            j = int.Parse(textBox2.Text);
+            double i, j;
+            i = double.Parse(textBox1.Text);
+            j = double.Parse(textBox2.Text);
 
             listBox1.Items.Add($"{i} * {j} = {i * j}");
         }
